config\feat(front): #46 upgrade astro
- upgrade astro
