+++ .gitignore
@@ -24,6 +24,6 @@ pnpm-debug.log*
 .idea/
 
 coverage/
-src/content/posts/artificial-vision.md
-src/content/posts/steps-by-step.md
+src/blog/steps-by-step.md
 public/el_problema_de_los_puentes.png
+src/blog/developing-secure-software.md
