+++ .vscode/settings.json
@@ -5,6 +5,7 @@
   },
   "[markdown]": {
     "editor.rulers": [80],
+    "editor.defaultFormatter": "esbenp.prettier-vscode",
     "diffEditor.ignoreTrimWhitespace": true
   },
   "[mdx]": {
