+++ astro.config.mjs
@@ -1,6 +1,5 @@
 // @ts-check
 import { defineConfig } from "astro/config";
-import mdx from "@astrojs/mdx";
 import sitemap from "@astrojs/sitemap";
 import rehypeExternalLinks from "rehype-external-links";
 
@@ -27,14 +26,5 @@ export default defineConfig({
       wrap: true,
     },
   },
-  integrations: [
-    mdx({
-      syntaxHighlight: "shiki",
-      shikiConfig: {
-        theme: "tokyo-night",
-        wrap: true,
-      },
-    }),
-    sitemap(),
-  ],
+  integrations: [sitemap()],
 });
