+++ project-words.txt
@@ -15,8 +15,12 @@ blockquote
 builtins
 Checksumming
 clearfix
+Saltzer
+bypassability
+Kubelet
 cowsay
 currentcolor
+BYOD
 cvzf
 datetime
 devaluación
@@ -31,10 +35,33 @@ focusring
 frontmatter
 functools
 getenv
+Pentesting
+mitigación
 gnugrep
 gnused
 gunicorn
+Metasploit
+Wireshark
+priorización
+mitigaciones
+CAPEC
+pentesters
+OSINT
+Escaneos
+remediación
+Priorización
+Exploitability
+Subscore
+Footprinting
+Mapeo
 hashchange
+nmap
+Vishing
+Smishing
+cybercriminals
+NGFW
+SIEM
+ciphertext
 hlsearch
 Introducción
 keyof
