+++ src/blog/artificial-vision.md
@@ -1,0 +1,55 @@
+---
+title: Artificial Vision
+pubDate: 2024-10-27
+description: "Artificial Vision"
+tags: ["artificial-vision"]
+isDraft: true
+snippet:
+  language: "bash"
+  code: "
+mkdir dir_name\n
+cd dir_name\n
+touch file_name\n
+echo 'foo' > file_name\n
+ls\n
+cat file_name\n
+rm file_name\n
+"
+---
+
+## Kind of Algorithms
+
+- Input Image -> Output Image: Image Processing.
+  - Image Processing: Image processing is a method to perform some operations
+    on an image, in order to get an enhanced image or to extract some useful
+    information from it. It is a type of signal processing in which input is an
+    image and output may be image or characteristics/features associated with
+    that image.
+- 3D input image -> 2D output image: Graphics computing
+  - Graphics computing: Graphics computing is a method to perform some
+    operations on a 3D image, in order to get a 2D image or to extract some useful
+    information from it. It is a type of signal processing in which input is a 3D
+    image and output may be a 2D image or characteristics/features associated with
+    that image.
+
+## Image Processing on the space domain
+
+We don't use transformations. Instead we ise pixels information as well as gray levels.
+
+### Geometric Processing
+
+- rotation.
+- Zoom.
+
+### Histogram Processing
+
+- Equalization of histogram.
+
+### Arithmetic Operations
+
+- Addition.
+- Multiplication.
+
+### Spatial filters
+
+- Convolution.
