+++ src/blog/wsgi-vs-asgi-protocols.md
@@ -6,7 +6,6 @@ web development. Learn key differences, implementation examples, and when
 to use each protocol for optimal web application performance."
 tags: [WSGI, ASGI, Python web development, web protocols, FastAPI, Flask,
 web servers, async programming, web sockets, HTTP/2]
-layout: "../../layouts/Post.astro"
 snippet:
   language: "python"
   code: "def app(environ, start_response):\n
