+++ src/components/Aside.astro
@@ -31,8 +31,8 @@ const { titles } = Astro.props;
       <a href={`${BASE_URL}about/`}
         ><Icon icon="about" size="35px" color="currentcolor" />about</a
       >
-      <a href={`${BASE_URL}posts/`}
-        ><Icon icon="article" size="35px" color="currentcolor" />posts</a
+      <a href={`${BASE_URL}blog/`}
+        ><Icon icon="article" size="35px" color="currentcolor" />blog</a
       >
       {
         titles.map((title) => (
