--- src/content/posts/solid-python.md
@@ -1,21 +1,0 @@
----
-title: SOLID Python
-pubDate: 2025-02-19
-description: "In this post, we will explore the SOLID principles in Python. From
-  Single Responsibility Principle (SRP), Open/Closed Principle (OCP), Liskov
-  Substitution Principle (LSP), Interface Segregation Principle (ISP),
-  and Dependency Inversion Principle (DIP). Let's dive in!"
-tags: ["programming", "python", "SOLID"]
-isDraft: true
-layout: "../../layouts/Post.astro"
-snippet:
-  language: "python"
-  code: "SOLID Python"
----
-
-## SRP: Single Responsibility Principle
-
-> A class should have only one reason to change.
-
-In other words, a class should have only one responsibility or job. This
-principle helps to keep classes small, focused, and maintainable.
