+++ src/layouts/Post.astro
@@ -1,16 +1,14 @@
 ---
+import type { CollectionEntry } from "astro:content";
 import "../styles/base.css";
 import Footer from "components/Footer.astro";
 import Head from "components/Head.astro";
 import Hero from "components/Hero.astro";
 import TableOfContent from "components/TableOfContent.astro";
-import type { IPostFrontmatter } from "content/config";
 import Aside from "components/Aside.astro";
 
 export interface Props {
-  file: string;
-  url: string;
-  frontmatter: IPostFrontmatter;
+  frontmatter: CollectionEntry<"blog">["data"];
   headings: { slug: string; text: string; depth: number }[];
 }
