+++ src/pages/blog/[...slug].astro
@@ -1,0 +1,17 @@
+---
+import Post from "layouts/Post.astro";
+import { getCollection, render } from "astro:content";
+export async function getStaticPaths() {
+  const posts = await getCollection("blog", ({ data }) => !data.isDraft);
+  return posts.map((post) => ({
+    params: { slug: post.id },
+    props: { post },
+  }));
+}
+const { post } = Astro.props;
+const { Content, headings } = await render(post);
+---
+
+<Post frontmatter={post.data} headings={headings}>
+  <Content />
+</Post>
