--- src/pages/posts/[...slug].astro
@@ -1,24 +1,0 @@
----
-import { getCollection } from "astro:content";
-
-export async function getStaticPaths() {
-  const posts = await getCollection("posts");
-
-  return posts.map((post) => {
-    return {
-      params: {
-        slug: post.slug,
-      },
-      props: {
-        ...post,
-        title: post.data.title,
-      },
-    };
-  });
-}
-
-const { render } = Astro.props;
-const { Content } = await render();
----
-
-<Content />
