--- src/pages/posts/index.astro
@@ -1,27 +1,0 @@
----
-import Base from "layouts/Base.astro";
-import PostPreview from "components/PostPreview.astro";
-import { getCollection } from "astro:content";
-
-const allPosts = (
-  await getCollection("posts", (entry) => !entry.data.isDraft)
-).sort((left, right) => {
-  return right?.data?.pubDate.valueOf() - left?.data?.pubDate.valueOf();
-});
----
-
-<Base title="Blog | Yeison Liscano" description="Yeison Liscano Blog">
-  {
-    allPosts.map((post) => (
-      <PostPreview
-        title={post.data.title}
-        pubDate={post.data.pubDate}
-        description={post.data.description}
-        tags={post.data.tags}
-        slug={post.slug}
-        readingTime={post.data?.readingTime}
-        snippet={post.data.snippet}
-      />
-    ))
-  }
-</Base>
